+++ test/augmenters/test_geometric.py
@@ -670,7 +670,7 @@ def test_Affine():
         img_aug_mask = img_aug > 255*0.1
         hm_aug_mask = hm_aug.arr_0to1 > 0.1
         same = np.sum(img_aug_mask == hm_aug_mask[:, :, 0])
-        assert (same / img_aug_mask.size) >= 0.99
+        assert (same / img_aug_mask.size) >= 0.95
 
         # measure alignment between images and heatmaps when rotating
         # here with smaller heatmaps