+++ CHANGELOG.md
@@ -58,6 +58,13 @@
 * Added property `Polygon.width`.
 * [mildly breaking] Changed the output of `Polygon.clip_out_of_image()` from `MultiPolygon` to `list` of `Polygon`.
   This breaks for anybody who has already used `Polygon.clip_out_of_image()`. 
+* Fixed `Affine` heatmap augmentation producing arrays with values outside the range `[0.0, 1.0]` when `order` was set to `3`.
+* Changed `Affine` to always use `order=3` for heatmap augmentation.
+* Fixed `PiecewiseAffine` heatmap augmentation producing arrays with values outside the range `[0.0, 1.0]` when `order` was set to `3`.
+* Changed `PiecewiseAffine` to always use `order=3` for heatmap augmentation.
+* Changed `ElasticTransformation` to always use `order=3` for heatmap augmentation.
+* Changed check in `HeatmapsOnImage` that validates whether the input array is within the desired value range `[min_value, max_value]` 
+  from a hard exception to a soft warning (with clipping). Also improved the error message a bit.
 
 
 # 0.2.8